I remove API call definitions from react component into util directory.
